fix: typescript defs for function properties
The existing Typescript defs weren't recognized as properties of the default export, so attempting to use them in a Typescript project cause an error.
add types for exec and create functions

As far as I can tell this is the only way to export the exec and create functions both as function properties and individual exports.